requirements.txt
Django==1.11.4
psycopg2==2.7.3
django-encrypted-id==0.2.0

